// src/types/global.d.ts
export interface Notification {
    timestamp?: number;
    id: number;
    type: "success" | "error" | "warning" | "default";
    title?: string;
    content: string;
}